Move the new semantic analyzer out of newsemanal (#7296)
