+++ mypy/semanal.py
@@ -90,7 +90,7 @@ from mypy.types import (
 )
 from mypy.type_visitor import TypeQuery
 from mypy.nodes import implicit_module_attrs
-from mypy.newsemanal.typeanal import (
+from mypy.typeanal import (
     TypeAnalyser, analyze_type_alias, no_subscript_builtin_alias,
     TypeVariableQuery, TypeVarList, remove_dups, has_any_from_unimported_type,
     check_for_explicit_any, type_constructors, fix_instance_types
@@ -103,13 +103,13 @@ from mypy.plugin import (
 )
 from mypy.util import correct_relative_import, unmangle, module_prefix
 from mypy.scope import Scope
-from mypy.newsemanal.semanal_shared import (
+from mypy.semanal_shared import (
     SemanticAnalyzerInterface, set_callable_name, calculate_tuple_fallback, PRIORITY_FALLBACKS
 )
-from mypy.newsemanal.semanal_namedtuple import NamedTupleAnalyzer
-from mypy.newsemanal.semanal_typeddict import TypedDictAnalyzer
-from mypy.newsemanal.semanal_enum import EnumCallAnalyzer
-from mypy.newsemanal.semanal_newtype import NewTypeAnalyzer
+from mypy.semanal_namedtuple import NamedTupleAnalyzer
+from mypy.semanal_typeddict import TypedDictAnalyzer
+from mypy.semanal_enum import EnumCallAnalyzer
+from mypy.semanal_newtype import NewTypeAnalyzer
 from mypy.reachability import (
     infer_reachability_of_if_statement, infer_condition_value, ALWAYS_FALSE, ALWAYS_TRUE,
     MYPY_TRUE, MYPY_FALSE
