+++ mypy/semanal_classprop.py
