+++ mypy/semanal_enum.py
@@ -10,7 +10,7 @@ from mypy.nodes import (
     UnicodeExpr, TupleExpr, ListExpr, DictExpr, Var, SymbolTableNode, MDEF, ARG_POS,
     EnumCallExpr
 )
-from mypy.newsemanal.semanal_shared import SemanticAnalyzerInterface
+from mypy.semanal_shared import SemanticAnalyzerInterface
 from mypy.options import Options
