+++ mypy/semanal_newtype.py
@@ -14,10 +14,10 @@ from mypy.nodes import (
     UnicodeExpr, Block, FuncDef, Argument, TypeInfo, Var, SymbolTableNode, MDEF, ARG_POS,
     PlaceholderNode
 )
-from mypy.newsemanal.semanal_shared import SemanticAnalyzerInterface
+from mypy.semanal_shared import SemanticAnalyzerInterface
 from mypy.options import Options
 from mypy.exprtotype import expr_to_unanalyzed_type, TypeTranslationError
-from mypy.newsemanal.typeanal import check_for_explicit_any, has_any_from_unimported_type
+from mypy.typeanal import check_for_explicit_any, has_any_from_unimported_type
 from mypy.messages import MessageBuilder, format_type
