+++ mypy/semanal_shared.py
