+++ mypy/semanal_typeddict.py
@@ -10,10 +10,10 @@ from mypy.nodes import (
     ClassDef, RefExpr, TypeInfo, AssignmentStmt, PassStmt, ExpressionStmt, EllipsisExpr, TempNode,
     DictExpr, ARG_POS, ARG_NAMED
 )
-from mypy.newsemanal.semanal_shared import SemanticAnalyzerInterface
+from mypy.semanal_shared import SemanticAnalyzerInterface
 from mypy.exprtotype import expr_to_unanalyzed_type, TypeTranslationError
 from mypy.options import Options
-from mypy.newsemanal.typeanal import check_for_explicit_any, has_any_from_unimported_type
+from mypy.typeanal import check_for_explicit_any, has_any_from_unimported_type
 from mypy.messages import MessageBuilder
 
 TPDICT_CLASS_ERROR = ('Invalid statement in TypedDict definition; '
