+++ mypy/server/aststrip.py
