+++ setup.py
@@ -169,7 +169,7 @@ setup(name='mypy',
       py_modules=[],
       ext_modules=ext_modules,
       packages=[
-          'mypy', 'mypy.test', 'mypy.server', 'mypy.plugins', 'mypy.newsemanal', 'mypy.dmypy'
+          'mypy', 'mypy.test', 'mypy.server', 'mypy.plugins', 'mypy.dmypy',
       ],
       package_data={'mypy': package_data},
       entry_points={'console_scripts': ['mypy=mypy.__main__:console_entry',
